move Beacon* structs
