+++ ethereum-common/src/lib.rs
@@ -36,3 +36,5 @@ pub use trie_db;
 pub const SLOTS_PER_EPOCH: u64 = 32;
 pub const EPOCHS_PER_SYNC_COMMITTEE: u64 = 256;
 pub const SYNC_COMMITTEE_SIZE: usize = 512;
+// https://github.com/ethereum/consensus-specs/blob/dev/specs/altair/light-client/p2p-interface.md#configuration
+pub const MAX_REQUEST_LIGHT_CLIENT_UPDATES: u8 = 128;
